Add files via upload
This is the first version of code that I used for my MPU 6050. It is a version of the code provided by the Instructable Project titled "Simple Arduino 3-Axis Accelerometer Datalogger." Because I did not need an SD Card, I prompted ChatGPT to modify the code so that it excluded the Sd Card but kept everything else the same.