Update Knee_brace_code.ino
To create this version of code, I prompted ChatGPT to include an on/off switch and a haptic since the end goal was to put the accelerometer on a knee brace. I also prompted ChatGPT to make it so that the haptic vibrated only when the x and y both reached a certain threshold (at this time I confused the up and down direction was the y direction). I chose random values because at this point I had not calculated the G's involved in a jump. At first, the haptic would vibrate if only one of the conditions was met even if the other was not. However, after much trial and error, ChatGPT produced this current version that worked as intended.