Update Knee_brace_code.ino
At this point I was testing, so I asked ChatGPT to generate code that only printed the y-value of acceleration if it was greater than the absolute value of 4G's (I later realized I needed to detect 5G's) because I was trying to measure the height I should lift the "leg" in order to produce the desired y-value (I realized that I hadn't confused the y and the z. Given the orientation of the accelerometer, the y-direction is up and down). However, the MPU 6050 did not read any values upon impact even after changing the range of values the accelerometer could read to +/-8G's. I then tried increasing the sampling rate, but that change just produced arbitrary values. At this point, I realized I needed an accelerometer that had a higher bandwidth so that it could detect the instantaneous G value upon impact.